Populate list with Header and Navbar comp wth type
