+++ src/components/header/header.css
@@ -12,6 +12,10 @@
   margin: 20px 0 100px 0;
 }
 
+.headerContainer.listMode {
+  margin: 20px 0 0 0;
+}
+
 .headerList {
   display: flex;
   gap: 40px;
