--- src/components/navbar/header/Header.jsx
@@ -1,190 +1,0 @@
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import './header.css';
-import {
-  faBed,
-  faCalendarDays,
-  faCar,
-  faPerson,
-  faPlane,
-  faTaxi,
-} from '@fortawesome/free-solid-svg-icons';
-import { useState } from 'react';
-import { DateRangePicker } from 'react-date-range';
-import { format } from 'date-fns';
-import 'react-date-range/dist/styles.css'; // main style file
-import 'react-date-range/dist/theme/default.css'; // theme css file
-
-const Header = () => {
-  const [openDate, setOpenDate] = useState(false);
-
-  const [date, setDate] = useState({
-    startDate: new Date(),
-    endDate: new Date(),
-    key: 'selection',
-  });
-
-  const handleSelect = (ranges) => {
-    setDate(ranges.selection);
-
-    console.log(ranges);
-  };
-
-  const formattedStartDate = format(date.startDate, 'MMM dd, yyyy');
-  const formattedEndDate = format(date.endDate, 'MMM dd, yyyy');
-  const formattedDateRange = `${formattedStartDate} - ${formattedEndDate}`;
-
-  const [openOptions, setOpenOptions] = useState(false);
-
-  const [options, setOptions] = useState({
-    adult: 1,
-    children: 0,
-    room: 1,
-  });
-
-  const handleOption = (name, operation) => {
-    setOptions((prev) => {
-      return {
-        ...prev,
-        [name]: operation === 'i' ? options[name] + 1 : options[name] - 1,
-      };
-    });
-  };
-
-  return (
-    <div className="header">
-      <div className="headerContainer">
-        <div className="headerList">
-          <div className="headerListItem active">
-            <FontAwesomeIcon icon={faBed} />
-            <span>Stays</span>
-          </div>
-          <div className="headerListItem">
-            <FontAwesomeIcon icon={faPlane} />
-            <span>Flight</span>
-          </div>
-          <div className="headerListItem">
-            <FontAwesomeIcon icon={faCar} />
-            <span>Car rentals</span>
-          </div>
-          <div className="headerListItem">
-            <FontAwesomeIcon icon={faBed} />
-            <span>Attractions</span>
-          </div>
-          <div className="headerListItem">
-            <FontAwesomeIcon icon={faTaxi} />
-            <span>Airport taxis</span>
-          </div>
-        </div>
-        <h1 className="headerTitle">A lifetime of discounts? It's Genius.</h1>
-        <p className="headerDesc">
-          Get rewarded for your travels - unlock instant savings of 10% or more
-          with a free nabeelbooking account
-        </p>
-        <button className="headerBtn">Sign in / Register</button>
-        <div className="headerSearch">
-          <div className="headerSearchItem">
-            <FontAwesomeIcon icon={faBed} className="headerIcon" />
-            <input
-              type="text"
-              placeholder="where are you going"
-              className="headerSearchInput"
-            />
-          </div>
-          <div className="headerSearchItem">
-            <FontAwesomeIcon icon={faCalendarDays} className="headerIcon" />
-            <span
-              onClick={() => setOpenDate(!openDate)}
-              className="headerSearchText"
-            >
-              {formattedDateRange}
-            </span>
-            {openDate && (
-              <DateRangePicker
-                ranges={[date]}
-                onChange={handleSelect}
-                className="date"
-              />
-            )}
-          </div>
-          <div className="headerSearchItem">
-            <FontAwesomeIcon icon={faPerson} className="headerIcon" />
-            <span
-              onClick={() => {
-                setOpenOptions(!openOptions);
-              }}
-              className="headerSearchText"
-            >{`${options.adult}- adult ${options.children}- children ${options.room} - room`}</span>
-            {openOptions && (
-              <div className="options">
-                <div className="optionItem">
-                  <span className="optionText">Adult</span>
-                  <div className="optionCounter">
-                    <button
-                      disabled={options.adult <= 1}
-                      onClick={() => handleOption('adult', 'd')}
-                      className="optionCounterButton"
-                    >
-                      -
-                    </button>
-                    <span className="optionCounterNumber">{options.adult}</span>
-                    <button
-                      onClick={() => handleOption('adult', 'i')}
-                      className="optionCounterButton"
-                    >
-                      +
-                    </button>
-                  </div>
-                </div>
-                <div className="optionItem">
-                  <span className="optionText">Children</span>
-                  <div className="optionCounter">
-                    <button
-                      disabled={options.children <= 0}
-                      className="optionCounterButton"
-                      onClick={() => handleOption('children', 'd')}
-                    >
-                      -
-                    </button>
-                    <span className="optionCounterNumber">
-                      {options.children}
-                    </span>
-                    <button
-                      className="optionCounterButton"
-                      onClick={() => handleOption('children', 'i')}
-                    >
-                      +
-                    </button>
-                  </div>
-                </div>
-                <div className="optionItem">
-                  <span className="optionText">Rooms</span>
-                  <div className="optionCounter">
-                    <button
-                      disabled={options.room <= 1}
-                      onClick={() => handleOption('room', 'd')}
-                      className="optionCounterButton"
-                    >
-                      -
-                    </button>
-                    <span className="optionCounterNumber">{options.room}</span>
-                    <button
-                      onClick={() => handleOption('room', 'i')}
-                      className="optionCounterButton"
-                    >
-                      +
-                    </button>
-                  </div>
-                </div>
-              </div>
-            )}
-          </div>
-          <div className="headerSearchItem">
-            <button className="headerBtn">Search</button>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export default Header;
