+++ src/pages/home/Home.jsx
@@ -1,5 +1,5 @@
+import Header from '../../components/header/Header';
 import Navbar from '../../components/navbar/Navbar';
-import Header from '../../components/navbar/header/Header';
 import './home.css';
 const Home = () => {
   return (
