+++ src/pages/list/List.jsx
@@ -1,9 +1,13 @@
-import React from 'react'
+import Header from '../../components/header/Header';
+import Navbar from '../../components/navbar/Navbar';
 
 const List = () => {
   return (
-    <div>List</div>
-  )
-}
+    <>
+      <Navbar />
+      <Header type="list" />
+    </>
+  );
+};
 
-export default List
+export default List;
